// src/app/api-services-v2/model/costCenterRequestDTO.ts
/**
 * Bestellsystem API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CostCenterRequestDTO { 
    id: string;
    name: string;
    begin_date?: string;
    end_date?: string;
    comment?: string;
}